extract simulations out of class
